Fix CI and tests and README. Improved the wheel generation (#1)

+++ src/pywhl.py
@@ -16,7 +16,6 @@ import shutil
 import site
 import string
 import sys
-import tempfile
 import zipfile
 from collections.abc import Generator
 from dataclasses import dataclass
@@ -91,7 +90,7 @@ def parse_manifest(manifest: Path) -> BuildInfo:
         return list(filter(len, filter(None, paths.split(";"))))
 
     def create_source_dest_paths(paths: list[str]) -> dict[Path, Path]:
-        return {Path(src): Path(dst or Path(src).name) for path in paths for src, dst in [(([*path.split("@"), ""])[0:2])]}
+        return {Path(src): Path(dst or Path(src).name) for path in paths for src, dst in [([*path.split("@"), ""])[:2]]}
 
     def parse_manifest(manifest: Path) -> ModuleInfo:
         modconfig = read_config(manifest)
@@ -111,7 +110,7 @@ def parse_manifest(manifest: Path) -> BuildInfo:
 
         return modinfo
 
-    def path_or_none(val: str| None)-> Path | None:
+    def path_or_none(val: str | None) -> Path | None:
         return None if val is None else Path(val)
 
     return BuildInfo(
@@ -154,54 +153,62 @@ class CMakeBuildWheel:
         if self.build_info.entry_points_file:
             self.dependencies.add(self.build_info.entry_points_file)
 
-    def _generate_editable_finder(self, name: str) -> str:  # noqa: C901
-        """Create a string containing the code for the MetaPathFinder and PathEntryFinder."""
+    def _script_path_mapping(self, mod: ModuleInfo, script_path: Path) -> dict[str, str]:
+        script_root = mod.src_root / script_path
+        if not script_root.exists():
+            raise CMakeBuildWheelError(f"Script {script_root.as_posix()} not found")
         mapping: dict[str, str] = {}
-        if len(self.build_info.modules) == 0:
+        if script_root.is_dir():
+            if (script_root / "__init__.py").exists():
+                mapping[mod.name] = script_root.as_posix()
+            else:
+                for subpath in script_root.rglob("*.py"):
+                    mapping[f"{mod.name}.{subpath.stem}"] = subpath.as_posix()
+        elif script_root.suffix == ".py":
+            mapping[f"{mod.name}.{script_root.stem}"] = script_root.as_posix()
+            if script_path.name == "__init__.py":
+                mapping[mod.name] = script_path.absolute().as_posix()
+        return mapping
+
+    def _build_editable_mapping(self) -> dict[str, str]:
+        if not self.build_info.modules:
             raise ValueError("Empty self.build_info.modules")
+        mapping: dict[str, str] = {}
         for mod in self.build_info.modules:
             for script_path in mod.scripts:
-                script_root = mod.src_root / script_path
-                if not script_root.exists():
-                    raise CMakeBuildWheelError(f"Script {script_root.as_posix()} not found")
-                if script_root.is_dir():
-                    if (script_root / "__init__.py").exists():
-                        mapping[mod.name] = script_root.as_posix()
-                    else:
-                        for subpath in script_root.rglob("*.py"):
-                            mapping[f"{mod.name}.{subpath.stem}"] = subpath.as_posix()
-                if script_root.is_file() and script_root.suffix == ".py":
-                    mapping[f"{mod.name}.{script_root.stem}"] = script_root.as_posix()
-                    if script_path.name == "__init__.py":
-                        mapping[mod.name] = script_path.absolute().as_posix()
-
+                mapping.update(self._script_path_mapping(mod, script_path))
             for target_lib in mod.target_libs:
-                libname = target_lib.stem
-                mapping[libname] = target_lib.as_posix()
-                mapping[re.sub(r"^lib", "", libname)] = target_lib.as_posix()
-                mapping[re.sub(r".abi[0-9\.]+", "", libname)] = target_lib.as_posix()
+                for libname in [
+                    target_lib.stem,
+                    re.sub(r".abi[0-9\.]+", "", target_lib.stem),
+                    re.sub(r"^lib", "", target_lib.stem),
+                ]:
+                    mapping[libname] = target_lib.as_posix()
+                    mapping[mod.name + "." + libname] = target_lib.as_posix()
             for target_bin in mod.target_bins:
                 mapping[target_bin.stem] = target_bin.as_posix()
+                mapping[mod.name + "." + target_bin.stem] = target_bin.as_posix()
             for data_path in mod.data:
                 mapping[data_path.name] = data_path.as_posix()
+        return mapping
+
+    def _generate_editable_finder(self, name: str) -> str:
+        """Create a string containing the code for the MetaPathFinder and PathEntryFinder."""
+        mapping = self._build_editable_mapping()
         namespaces: dict[str, str] = {}
         tmpl = (Path(__file__).parent / "editable_finder.py").read_text()
         tmpl = tmpl.replace(
             "MAPPING: dict[str, str] = {}  # TEMPLATE-SUBSTITUTION-MARKER",
             f"MAPPING: dict[str, str] = {mapping!r}",
-            count=1,
         )
         tmpl = tmpl.replace(
             "NAMESPACES: dict[str, str] = {}  # TEMPLATE-SUBSTITUTION-MARKER",
             f"NAMESPACES: dict[str, str] = {namespaces!r}",
-            count=1,
         )
-        tmpl = tmpl.replace(
+        return tmpl.replace(
             'PATH_PLACEHOLDER: str = ".__path_hook__"  # TEMPLATE-SUBSTITUTION-MARKER',
             f'PATH_PLACEHOLDER: str = {name!r} + ".__path_hook__"',
-            count=1,
         )
-        return tmpl
 
     def get_requires_for_build_sdist(self, _config_settings: dict[str, object] | None = None) -> list[str]:
         return []  # No dependencies for building with this file as the build backend
@@ -396,14 +403,15 @@ class CMakeBuildWheel:
         """
         raise CMakeBuildWheelError("build_editable Untested")
 
-    def _build_editable_at(self, out_dir: Path) -> None:
+    def _build_editable_at(self, out_dir: Path, _force_reinstall: bool) -> None:
         for fpath, src_file, contents in self._generate_wheel_content(editable=True):
             fabspath = out_dir / fpath
             fabspath.parent.mkdir(parents=True, exist_ok=True)
             if src_file is not None:
                 _ = shutil.copyfile(src_file, fabspath)
             elif contents is not None:
-                _ = fabspath.write_text(contents, encoding="utf-8")
+                if not fabspath.exists() or fabspath.read_text(encoding="utf-8") != contents:
+                    _ = fabspath.write_text(contents, encoding="utf-8")
             else:
                 raise ValueError(f"Neither file nor contents found for {fpath.as_posix()}")
 
@@ -436,20 +444,9 @@ class CMakeBuildWheel:
 
     def install(self, editable: bool, force_reinstall: bool = False) -> None:
         if editable:
-            with tempfile.TemporaryDirectory() as tmpdir:
-                self._build_editable_at(Path(tmpdir))
-                files = [fpath.relative_to(tmpdir) for fpath in Path(tmpdir).rglob("*") if fpath.is_file()]
-                # The first one in site.getsitemodules() is the directory for the venv if we are in a venv
-                site_package_path = Path(site.getsitepackages()[0])
-                conflicts = [fpath for fpath in files if (site_package_path / fpath).exists()]
-                if not force_reinstall and len(conflicts) > 0:
-                    raise CMakeBuildWheelError(
-                        "Conflicting files: \n\t" + "\n\t".join(fpath.as_posix() for fpath in files),
-                    )
-                for fpath in files:
-                    (site_package_path / fpath).parent.mkdir(parents=True, exist_ok=True)
-                    _ = shutil.copy2(tmpdir / fpath, site_package_path / fpath)
-                    log.debug(f"Installed {(site_package_path / fpath).as_posix()}")
+            # The first one in site.getsitemodules() is the directory for the venv if we are in a venv
+            site_package_path = Path(site.getsitepackages()[0])
+            self._build_editable_at(Path(site_package_path), force_reinstall)
         else:
             raise CMakeBuildWheelError("Python install ('install') Unsupported")
